Refactor: rename `unleash` -> `feature-evaluator`
